DIV operation

+++ src/operations.h
@@ -11,7 +11,7 @@ double fixed64ToFloat(int64_t num);
 
 int addOp32(int32_t* num1, int32_t num2);
 int addOp64(int64_t* num1, int64_t num2);
-int divOp(int dividend, int divisor); 
+int32_t divOp(int64_t dividend, int32_t divisor); 
 int64_t convertSign64(int64_t num);
 int32_t convertSign32(int32_t num);
